+++ ext/node/lib.rs
@@ -439,8 +439,9 @@ deno_core::extension!(deno_node,
     "path/_constants.ts",
     "path/_interface.ts",
     "path/_util.ts",
+    "path/_posix.ts",
+    "path/_win32.ts",
     "path/common.ts",
-    "path/glob.ts",
     "path/mod.ts",
     "path/posix.ts",
     "path/separator.ts",
