+++ ext/node/polyfills/path.ts
@@ -1,4 +1,4 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
 export * from "ext:deno_node/path/mod.ts";
-import * as m from "ext:deno_node/path/mod.ts";
-export default { ...m };
+import m from "ext:deno_node/path/mod.ts";
+export default m;
