+++ ext/node/polyfills/path/mod.ts
@@ -3,13 +3,25 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
 
 import { isWindows } from "ext:deno_node/_util/os.ts";
-import _win32 from "ext:deno_node/path/win32.ts";
-import _posix from "ext:deno_node/path/posix.ts";
+import _win32 from "ext:deno_node/path/_win32.ts";
+import _posix from "ext:deno_node/path/_posix.ts";
 
-const path = isWindows ? _win32 : _posix;
+export const win32 = {
+  ..._win32,
+  win32: null as unknown as typeof _win32,
+  posix: null as unknown as typeof _posix,
+};
 
-export const win32 = _win32;
-export const posix = _posix;
+export const posix = {
+  ..._posix,
+  win32: null as unknown as typeof _win32,
+  posix: null as unknown as typeof _posix,
+};
+
+posix.win32 = win32.win32 = win32;
+posix.posix = win32.posix = posix;
+
+const path = isWindows ? win32 : posix;
 export const {
   basename,
   delimiter,
@@ -27,8 +39,6 @@ export const {
   toFileUrl,
   toNamespacedPath,
 } = path;
-
+export default path;
 export * from "ext:deno_node/path/common.ts";
-export { SEP, SEP_PATTERN } from "ext:deno_node/path/separator.ts";
 export * from "ext:deno_node/path/_interface.ts";
-export * from "ext:deno_node/path/glob.ts";
